Updates Conekta PHP library and fixes PHP 8.4 issues (#65)
* Points to PHP generator branch

Updates dependency to use the generator branch.

This allows for testing with PHP 8.4.

* Updates module to version 5.3.3

Increments the module version across all relevant files.

This update ensures consistency and reflects the latest changes in the module.

* Addresses PHP 8.4 compatibility issues

Ensures compatibility with PHP 8.4 by:

- Adding return type declarations to methods.
- Improving exception handling.
- Updating logger to handle errors effectively.

* Fixes PHP 8.4 compatibility issues

Addresses compatibility issues with PHP 8.4.

- Declares type hints for function arguments to improve code reliability.
- Updates property type declarations to ensure proper type handling.
- Removes obsolete code related to payment methods.

* Updates Conekta PHP library version

Updates the Conekta PHP library dependency to a stable release.

Fixes an issue where the module enablement command in the README was case-sensitive.

Ensures compatibility with PHP 8.4.

* Improves error handling in GatewayCommand by providing a default message for declined transactions